services/viewer/requirements.txt
gradio>=4.32.0
lancedb>=0.21.0
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
